% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{allocation_rule}
\alias{allocation_rule}
\alias{new_rule}
\alias{rule}
\title{Create a new treatment allocation rule}
\usage{
new_rule(parameters = list(), state = list(), class = character())
}
\arguments{
\item{parameters}{List of user-supplied parameters for the rule.}

\item{state}{List of values the rule keeps track of over evaluations.}

\item{class}{Name of the rule subclass to create.}
}
\description{
\code{new_rule()} is used as a base to create a new treatment allocation rule
subclass. Define a \code{\link[=rule_evaluate]{rule_evaluate()}} method to implement functionality.
The package's allocation rules are listed in the "See Also" section.
}
\seealso{
Other allocation rules: 
\code{\link{arjas_gasbarra}()},
\code{\link{random_allocation}()},
\code{\link{randomization_list}()},
\code{\link{thompson}()}
}
\concept{allocation rules}
