% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arjas_gasbarra.R
\name{arjas_gasbarra}
\alias{arjas_gasbarra}
\alias{rule_1}
\alias{rule_2}
\title{Allocate according to Arjas & Gasbarra's rule}
\usage{
arjas_gasbarra(
  epsilon,
  epsilon_1,
  epsilon_2,
  theta_low,
  delta,
  randomization_list = NULL
)

rule_1(epsilon, delta = 0, randomization_list = NULL)

rule_2(
  epsilon,
  epsilon_1,
  epsilon_2,
  theta_low,
  delta = 0,
  randomization_list = NULL
)
}
\description{
\code{arjas_gasbarra()} allows defining the general form introduced in Arjas &
Gasbarra (2021). \code{rule_1()} and \code{rule_2()} provide convenient shortcuts to
match the rules presented in the paper.
}
\seealso{
Other allocation rules: 
\code{\link{allocation_rule}},
\code{\link{random_allocation}()},
\code{\link{randomization_list}()},
\code{\link{thompson}()}
}
\concept{allocation rules}
