% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat-package.R
\docType{package}
\name{bat-package}
\alias{bat}
\alias{bat-package}
\title{bat: Bayesian Adaptive Treatment Allocation and Selection}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikko.marttila@orionpharma.com}

Other contributors:
\itemize{
  \item Orion Corporation [copyright holder]
}

}
\keyword{internal}
