% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.R
\name{phase}
\alias{phase}
\title{Define a study phase}
\usage{
phase(n, rule)
}
\arguments{
\item{n}{Number of subjects to accrue before the phase ends.}

\item{rule}{A \link{rule} for determining treatment allocation.}
}
\description{
Define a study phase
}
\seealso{
\code{\link[=simulate_phase]{simulate_phase()}} for generating simulated data for a phase.
}
