% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{random_allocation}
\alias{random_allocation}
\title{Allocate randomly with a given probability}
\usage{
random_allocation(p = NULL)
}
\description{
Allocate randomly with a given probability
}
\seealso{
Other allocation rules: 
\code{\link{allocation_rule}},
\code{\link{arjas_gasbarra}()},
\code{\link{randomization_list}()},
\code{\link{thompson}()}
}
\concept{allocation rules}
