% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{randomization_list}
\alias{randomization_list}
\title{Allocate by a pre-defined randomization list}
\usage{
randomization_list(randomization_list = NULL)
}
\description{
Allocate by a pre-defined randomization list
}
\seealso{
Other allocation rules: 
\code{\link{allocation_rule}},
\code{\link{arjas_gasbarra}()},
\code{\link{random_allocation}()},
\code{\link{thompson}()}
}
\concept{allocation rules}
