% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_phase}
\alias{simulate_phase}
\title{Simulate a study phase}
\usage{
simulate_phase(phase, theta, a, b, n_draws = 3000)
}
\arguments{
\item{phase}{A \link{phase} defining the number of subjects to accrue and the
treatment allocation rule to use.}

\item{theta}{A vector of true event probabilities in each arm.}

\item{a, b}{Vectors of parameters to the beta prior with elements for each
arm. Recycled to match the length of \code{theta} if possible.}

\item{n_draws}{The number of draws to sample from the posterior for
evaluating posterior probabilities in allocation rules.}
}
\description{
Simulate a study phase
}
\examples{
simulate_phase(
  phase(n = 240, rule = random_allocation(p = 1:2)),
  a = 1,
  b = 1,
  theta = c(0.3, 0.5)
)
}
\seealso{
\code{\link[=simulate_study]{simulate_study()}} for simulating a multi-phase study.
}
