% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study}
\alias{study}
\alias{single_phase_study}
\alias{multi_phase_study}
\title{Define a study}
\usage{
single_phase_study(...)

multi_phase_study(phases = list())
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=phase]{phase}}
  \describe{
    \item{\code{n}}{Number of subjects to accrue before the phase ends.}
    \item{\code{rule}}{A \link{rule} for determining treatment allocation.}
  }}

\item{phases}{A list of \link{phase}s that form the study.}
}
\description{
Define a study
}
\seealso{
\code{\link[=simulate_study]{simulate_study()}} for running a simulation of a study.
}
