% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{thompson}
\alias{thompson}
\title{Allocate according to Thompson's rule}
\usage{
thompson(kappa)
}
\description{
The randomization probability for each arm is determined by raising the
posterior probability of that arm being the best studied arm to the power
\code{kappa} given as a parameter.
}
\seealso{
Other allocation rules: 
\code{\link{arjas_gasbarra}()},
\code{\link{random_allocation}()},
\code{\link{randomization_list}()}
}
\concept{allocation rules}
