% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barts-package.R
\docType{package}
\name{barts-package}
\alias{barts}
\alias{barts-package}
\title{barts: Bayesian Adaptive Treatment Allocation and Selection}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikko.marttila@orionpharma.com}

Authors:
\itemize{
  \item Elja Arjas
  \item Dario Gasbarra
}

Other contributors:
\itemize{
  \item Orion Corporation [copyright holder]
}

}
\keyword{internal}
