% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_study}
\alias{simulate_study}
\title{Simulate a study}
\usage{
simulate_study(phases, theta, a, b, n_draws = 3000)
}
\arguments{
\item{phases}{A list of study \link{phase}s to simulate in sequence.}

\item{theta}{A vector of true event probabilities in each arm.}

\item{a}{Vectors of parameters to the beta prior with elements for each
arm. Recycled to match the length of \code{theta} if possible.}

\item{b}{Vectors of parameters to the beta prior with elements for each
arm. Recycled to match the length of \code{theta} if possible.}

\item{n_draws}{The number of draws to sample from the posterior for
evaluating posterior probabilities in allocation rules.}
}
\description{
Simulate a study
}
\examples{
simulate_study(
  phases = list(
    burn_in  = phase(n = 40, rule = randomization_list()),
    adaptive = phase(n = 240, rule = rule_1(epsilon = 0.1))
  ),
  a = 1,
  b = 1,
  theta = 0.2 + 1:4 / 10
)
}
\seealso{
\code{\link[=simulate_phase]{simulate_phase()}} for simulating a single phase.
}
