% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_study.R
\name{simulate_study}
\alias{simulate_study}
\title{Simulate a study}
\usage{
simulate_study(n_max, theta, a, b, rule, n_draws = 3000)
}
\arguments{
\item{n_max}{The number of subjects to stop after.}

\item{theta}{A vector of true event probabilities in each arm.}

\item{a}{A vector of "a" parameters for the beta prior.}

\item{b}{A vector of "b" parameters for the beta prior.}

\item{rule}{Treatment allocation rule to use.}

\item{n_draws}{The number of draws to sample from the posterior for
evaluating posterior probabilities.}
}
\description{
Simulate a study
}
\examples{
result <- simulate_study(
  n_max = 240,
  theta = 0.2 + 1:4 / 10,
  a = rep_len(1, 4),
  b = rep_len(1, 4),
  rule = rule_1(epsilon = 0.1)
)
}
